Allow statusbar items to be launch without the jank [1/2]
- Disable settings button

This is a combo of two commits that were pushed back in M, changed a few things but nothing major

https://github.com/DirtyUnicorns/android_frameworks_base/commit/e03d5f7e3a82aa064fcd51a60294bb8fd22137ec

https://github.com/DirtyUnicorns/android_frameworks_base/commit/b895182c71713d090570d866d9695aad51ea2b35
