+++ packages/SystemUI/res/xml/tuner_prefs.xml
@@ -16,16 +16,7 @@
 
 <PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:sysui="http://schemas.android.com/apk/res-auto"
-    android:title="@string/system_ui_tuner">
-
-    <com.android.systemui.tuner.TunerSwitch
-        android:key="qs_show_brightness"
-        android:title="@string/show_brightness"
-        sysui:defValue="true" />
-
-    <PreferenceScreen
-        android:key="status_bar"
-        android:title="@string/status_bar" >
+    android:title="@string/systemui_tuner_statusbar_title">
 
         <com.android.systemui.tuner.StatusBarSwitch
             android:key="rotate"
@@ -39,10 +30,6 @@
             android:key="managed_profile"
             android:title="@string/status_bar_work" />
 
-        <!-- ime -->
-        <!-- sync_failing -->
-        <!-- sync_active -->
-
         <com.android.systemui.tuner.StatusBarSwitch
             android:key="cast"
             android:title="@string/quick_settings_cast_title" />
@@ -55,15 +42,9 @@
             android:key="bluetooth"
             android:title="@string/quick_settings_bluetooth_label" />
 
-        <!-- nfc -->
-        <!-- tty -->
-        <!-- speakerphone -->
-
         <com.android.systemui.tuner.StatusBarSwitch
             android:key="zen"
             android:title="@string/quick_settings_dnd_label" />
-
-        <!-- mute -->
 
         <com.android.systemui.tuner.StatusBarSwitch
             android:key="volume"
@@ -85,69 +66,8 @@
             android:key="airplane"
             android:title="@string/status_bar_airplane" />
 
-        <!-- other weird signal stuff -->
-
-        <com.android.systemui.tuner.BatteryPreference
-            android:title="@string/battery"
-            android:summary="%s"
-            android:entries="@array/battery_options" />
-
         <com.android.systemui.tuner.StatusBarSwitch
             android:key="alarm_clock"
             android:title="@string/status_bar_alarm" />
-
-        <!-- secure -->
-
-        <com.android.systemui.tuner.ClockPreference
-            android:title="@string/tuner_time"
-            android:summary="%s"
-            android:entries="@array/clock_options" />
-
-    </PreferenceScreen>
-
-    <!--
-    <Preference
-        android:key="color_transform"
-        android:title="@string/color_and_appearance"
-        android:fragment="com.android.systemui.tuner.ColorAndAppearanceFragment" />
-    -->
-
-    <PreferenceScreen
-        android:key="volume_and_do_not_disturb"
-        android:title="@string/volume_and_do_not_disturb">
-
-        <!-- Action for this is
-             MetricsConstants.ACTION_TUNER_DO_NOT_DISTURB_VOLUME_PANEL -->
-        <com.android.systemui.tuner.TunerSwitch
-            android:key="sysui_show_full_zen"
-            android:title="@string/tuner_full_zen_title"
-            sysui:metricsAction="314" />
-
-        <!-- Action for this is
-             MetricsConstants.ACTION_TUNER_DO_NOT_DISTURB_VOLUME_SHORTCUT -->
-        <com.android.systemui.tuner.TunerSwitch
-            android:key="sysui_volume_down_silent,sysui_volume_up_silent"
-            android:title="@string/volume_dnd_silent"
-            sysui:defValue="true"
-            sysui:metricsAction="315" />
-
-    </PreferenceScreen>
-
-    <!--
-    <Preference
-        android:key="nav_bar"
-        android:title="@string/nav_bar"
-        android:fragment="com.android.systemui.tuner.NavBarTuner" />
-    -->
-
-    <Preference
-            android:key="other"
-            android:title="@string/other"
-            android:fragment="com.android.systemui.tuner.OtherPrefs" />
-
-    <!-- Warning, this goes last. -->
-    <Preference
-        android:summary="@string/tuner_persistent_warning"
-        android:selectable="false" />
 
 </PreferenceScreen>
